Move all experiment-related stuff to separate module
