+++ apps/class-solid/src/lib/store.ts
@@ -1,0 +1,4 @@
+import { createStore } from "solid-js/store";
+import type { Experiment } from "~/components/Experiment";
+
+export const [experiments, setExperiments] = createStore<Experiment[]>([]);
